requirements.txt
arabic-reshaper==3.0.0
asn1crypto==1.5.1
bcrypt==4.1.3
blinker==1.8.2
bson==0.5.10
certifi==2024.7.4
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.8
cssselect2==0.7.0
dnspython==2.6.1
Flask==3.0.3
gunicorn==22.0.0
html5lib==1.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==5.2.2
MarkupSafe==2.1.5
mysql-connector-python==8.4.0
oscrypto==1.3.0
packaging==24.1
pillow==10.4.0
pycparser==2.22
pyHanko==0.25.0
pyhanko-certvalidator==0.26.3
pymongo==4.8.0
PyMySQL==1.1.1
pypdf==4.2.0
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
qrcode==7.4.2
reportlab==4.2.2
requests==2.32.3
six==1.16.0
svglib==1.5.1
tinycss2==1.3.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uritools==4.0.3
urllib3==2.2.2
webencodings==0.5.1
Werkzeug==3.0.3
xhtml2pdf==0.2.16
